feat:SGED-51: Verify user id is same as auth ui for update and delete request

+++ src/controllers/userController.ts
@@ -45,6 +45,16 @@ export const createUser = async (req: Request, res: Response) => {
         USER_STATE: state
     } = req.body;
 
+    if (id != req.uid) {
+        const serverResponse = {
+            message: 'Error creating user',
+            error: 'You are not authorized to perform this action.'
+        };
+
+        e(serverResponse.message, new Error(serverResponse.error));
+        return res.status(403).json(serverResponse);
+    }
+
     if (!id || !firstName || !lastName || !email || !schoolLevel || !state) {
         const serverResponse = {
             message: 'Error creating user',
@@ -113,6 +123,16 @@ export const updateUser = async (req: Request, res: Response) => {
         USER_TESTER: tester
     } = req.body;
 
+    if (id != req.uid) {
+        const serverResponse = {
+            message: 'Error updating user',
+            error: 'You are not authorized to perform this action.'
+        };
+
+        e(serverResponse.message, new Error(serverResponse.error));
+        return res.status(403).json(serverResponse);
+    }
+
     if (
         id == null ||
         firstName == null ||
@@ -175,6 +195,16 @@ export const updateProfilePictureFileName = async (req: Request, res: Response) 
         USER_PROFILE_PICTURE_FILE_NAME: profilePictureFileName
     } = req.body;
 
+    if(userId != req.uid) {
+        const serverResponse = {
+            message: 'Error updating profile picture file name',
+            error: `You are not authorized to perform this action.`
+        };
+
+        e(serverResponse.message, new Error(serverResponse.error));
+        return res.status(403).json(serverResponse);
+    }
+
     if(!profilePictureFileName && !userId) {
         const serverResponse = {
             message: 'Error updating profile picture file name',
@@ -204,6 +234,16 @@ export const updateProfilePictureFileName = async (req: Request, res: Response) 
 
 export const deleteUser = async (req: Request, res: Response) => {
     const userId = req.params.userId;
+
+    if(userId != req.uid) {
+        const serverResponse = {
+            message: 'Error deleting user',
+            error: 'You are not authorized to perform this action.'
+        };
+
+        e(serverResponse.message, new Error(serverResponse.error));
+        return res.status(403).json(serverResponse);
+    }
 
     try {
         await userRepository.deleteUser(userId);
